pyproject.toml
[project]
name = "basic-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "fastapi[standard]>=0.116.1",
    "isort>=6.0.1",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.7",
    "scalar-fastapi>=1.2.2",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]
